#63 - libcore: no sysinfo for macos
