+++ README.md
@@ -174,9 +174,20 @@ Note: the installation mechanism does not remove the installation directory befo
 
 ------
 
-As of 0.0.40, the *libcore* runtime supports conditional compilation of a variety of features. See `Cargo.toml` for the features available in your code base.
+As of 0.0.40, the *libcore* runtime supports conditional compilation of a variety of features. 
 
-The *sysinfo* feature is currently not functional for *macos* builds, and is disabled by default.
+Currently supported features by namespace:
+
+```
+ default = [ "std", "sysinfo", "nix" ]
+ 
+ nix:     uname
+ std:     command, exit
+ sysinfo: sysinfo
+ 
+```
+
+The *sysinfo* feature is disabled on *macos* builds.
