+++ src/libcore/features/mod.rs
@@ -19,8 +19,6 @@ use crate::features::nix::nix_::{Core as _, Nix};
 use crate::features::std::std_::{Core as _, Std};
 #[cfg(feature = "sysinfo")]
 use crate::features::sysinfo::sysinfo_::{Core as _, Sysinfo};
-#[cfg(feature = "uname")]
-use crate::features::uname::uname_::{Core as _, Uname};
 
 #[cfg(feature = "nix")]
 pub mod nix;
@@ -28,8 +26,6 @@ pub mod nix;
 pub mod std;
 #[cfg(feature = "sysinfo")]
 pub mod sysinfo;
-#[cfg(feature = "uname")]
-pub mod uname;
 
 pub struct Feature {
     symbols: Vec<(&'static str, u16, CoreFunction)>,
@@ -37,7 +33,7 @@ pub struct Feature {
 }
 
 impl Feature {
-    fn install_feature(mu: &Mu, feature: Feature) -> Tag {
+    fn install(mu: &Mu, feature: Feature) -> Tag {
         let ns = Symbol::keyword(&feature.namespace);
         match Namespace::add_ns(mu, ns) {
             Ok(_) => (),
@@ -51,21 +47,19 @@ impl Feature {
 }
 
 pub trait Core {
-    fn add_features(_: &Mu) -> Vec<Tag>;
+    fn install_features(_: &Mu) -> Vec<Tag>;
 }
 
 impl Core for Feature {
-    fn add_features(_mu: &Mu) -> Vec<Tag> {
+    fn install_features(_mu: &Mu) -> Vec<Tag> {
         #[allow(clippy::let_and_return)]
         let features = vec![
             #[cfg(feature = "nix")]
-            Self::install_feature(_mu, Nix::make_feature(_mu)),
+            Self::install(_mu, Nix::make_feature(_mu)),
             #[cfg(feature = "std")]
-            Self::install_feature(_mu, Std::make_feature(_mu)),
+            Self::install(_mu, Std::make_feature(_mu)),
             #[cfg(feature = "sysinfo")]
-            Self::install_feature(_mu, Sysinfo::make_feature(_mu)),
-            #[cfg(feature = "uname")]
-            Self::install_feature(_mu, Uname::make_feature(_mu)),
+            Self::install(_mu, Sysinfo::make_feature(_mu)),
         ];
 
         features
