--- src/libcore/features/uname/mod.rs
@@ -1,1 +1,0 @@
-pub mod uname_;
